Add multiresolver (#605)
Multiresolver allows to resolve a name using multiple resolver chains.
Chains are selected using the TLD of the name (if no TLD is present, we use the default).
A chain resolves a name by consecutively calling Resolve on each element.
The name is resolved when the first resolver succeeds. It fails if all resolvers fail.