fix: eliminate local transaction nonce cache
Stop keeping the next nonce in our local state. Instead we always
calculate it from the blockchain nonce and the local pending
transactions.

Fixes issue #4826 where the local nonce cache was out of sync with the
blockchain for an unknown reason and because of that no transactions
got incorporated into the blockchain. This made unstaking or
participation in the reserve incentives impossible.

Tests adjusted by @martinconic.